requirements.txt
anthropic
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
requests
streamlit
pandas